# Conflicts:
#	.github/workflows/ci.yml
#	CHANGELOG.md
#	Makefile
#	docs/benchmark-protocol.md
#	pyproject.toml
#	src/decision_agent_bench/__init__.py

+++ src/decision_agent_bench/simulator/generator.py
@@ -265,7 +265,7 @@ def _populate(connection: sqlite3.Connection, config: GenerationConfig) -> None:
                     * seasonality
                     * decline
                 )
-                units = max(0, int(round(expected)) + rng.choice((-1, 0, 0, 0, 1)))
+                units = max(0, round(expected) + rng.choice((-1, 0, 0, 0, 1)))
                 if units == 0:
                     continue
                 product_id = product[0]
